Create activity if parentId contains valid data
The call to NServiceBusActivitySource.ActivitySource.StartActivity
returns null if no listeners are registered.

This change makes it so that an activity is created if the parentId
header contains valid data for the creation of ActivityContext.

It also sets the parent based on TraceId and SpanId, making the created
activity belong to the correct trace etc.